Temporarily disable weekly exhaustive measurements
We've received a few complaints regarding excessive
transmission of ICMP packets to specific prefixes. All signs
suggest it is the weekly exhaustive measurements sending
these probes but we do not understand the cause of the
sudden surge.  Therefore, we're temporarily suspending these
measurements until we gain clarity on the situation and
resolve the issue.